pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "max-quest-bot"
version = "0.1.0"
description = "MAX bot for quest walkthrough MVP"
readme = "BOT_README.md"
requires-python = ">=3.11"
dependencies = [
  "maxapi[fastapi]>=0.9.18,<1.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "python-multipart>=0.0.20,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[tool.setuptools.packages.find]
include = ["admin_panel*", "max_quest_bot*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
